+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Add detection of CachyOS. Closes #341.
+
 ## [5.0.0 (2025-11-03)](https://github.com/kdeldycke/extra-platforms/compare/v4.1.1...v5.0.0)
 
 - Relax dependencies to support Python 3.10.